requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
psycopg==3.2.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
